Memoir Preview Add Level & Exp

+++ chageunchageun/src/main/java/com/chageunchageun/chageunchageun/data/dto/User/UserDTO.java
@@ -21,6 +21,6 @@ public class UserDTO {
 
     private int level;
 
-    private int exp;
+    private float exp;
 
 }
